+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Download current and previous versions from [https://github.com/ludus-leonis/Fre
 
 * fixed no grid visible in low quality mode
 * fixed empty goblin label in RPG template
+* fixed hardcoded engine versions in integration tests
 
 ## v0.16.0 - Selecting Seal